+++ client/src/components/List/List.stories.js
@@ -1,0 +1,12 @@
+import React from "react";
+import List from "./List";
+
+export default {
+  title: "List",
+  component: List,
+};
+
+const Template = (args) => <List {...args} />;
+
+export const ListAllItems = Template.bind({});
+ListAllItems.args = <List />;
